move feed generation function
